% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_data}
\alias{check_data}
\title{Check the validity of the data structure}
\usage{
check_data(data, content_group_id)
}
\arguments{
\item{data}{dataframe}

\item{content_group_id}{Code that identifies the content}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of the data structure
}
\examples{
check_data(mtcars, "AS-HOSP")
}
