% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_content.R
\name{check_data_content}
\alias{check_data_content}
\title{Check the content of your data's variables}
\usage{
check_data_content(data, content_group_id)
}
\arguments{
\item{data}{dataframe of wrangled data}

\item{content_group_id}{Code that identifies the content}
}
\value{
list of exit statuses for each variable
}
\description{
Check the content of your data's variables
}
\examples{
data <-
 mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))

check_data_content(data, "AS-HOSP")
}
