% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_jurisdiction_code}
\alias{check_jurisdiction_code}
\title{Check the validity of a jurisdiction code}
\usage{
check_jurisdiction_code(jurisdiction_code)
}
\arguments{
\item{jurisdiction_code}{Two-letter state code}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of a jurisdiction code
}
\examples{
check_jurisdiction_code("CA")
}
