% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_mcn}
\alias{check_mcn}
\title{Check validity of a Metadata Control Number}
\usage{
check_mcn(mcn)
}
\arguments{
\item{mcn}{Metadata Control Number provided by EPHT}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check validity of a Metadata Control Number
}
\examples{
check_mcn("12345678-1234-1234-1234-123456789012")
}
