% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_state_fips_code}
\alias{check_state_fips_code}
\title{Check the validity of a state FIPS code}
\usage{
check_state_fips_code(state_fips_code)
}
\arguments{
\item{state_fips_code}{FIPS code of the state}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of a state FIPS code
}
\examples{
check_state_fips_code("06")
}
