% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_submitter_email}
\alias{check_submitter_email}
\title{Check the validity of a submitter email}
\usage{
check_submitter_email(submitter_email)
}
\arguments{
\item{submitter_email}{email address string}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of a submitter email
}
\examples{
check_submitter_email("myemail@email.com")
}
