% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_submitter_name}
\alias{check_submitter_name}
\title{Check the validity of a submitter name}
\usage{
check_submitter_name(submitter_name)
}
\arguments{
\item{submitter_name}{First and last name of person submitting data}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of a submitter name
}
\examples{
check_submitter_name("Firstname Lastname")
}
