% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_submitter_title}
\alias{check_submitter_title}
\title{Check the validity of a submitter title}
\usage{
check_submitter_title(submitter_title)
}
\arguments{
\item{submitter_title}{Job title of person submitting data}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the validity of a submitter title
}
\examples{
check_submitter_title("Data Scientist")
}
