% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{collapse_race}
\alias{collapse_race}
\title{Collapse Race based on EPHT logic}
\usage{
collapse_race(race, na_is_unknown = TRUE)
}
\arguments{
\item{race}{Race string}

\item{na_is_unknown}{Convert NAs to "U" or leave as-is}
}
\value{
Race string or NA_character_
}
\description{
Collapse Race based on EPHT logic
}
\examples{
collapse_race("Asian")
}
