% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distilleR-package.R
\docType{package}
\name{distilleR-package}
\alias{distilleR}
\alias{distilleR-package}
\title{distilleR: A Wrap Around the DistillerSR APIs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The distilleR package provides a pool of functions to query DistillerSR through its APIs. It features authentication and utilities to retrieve data from DistillerSR projects and reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openefsa.github.io/distilleR/}
  \item Report bugs at \url{https://github.com/openefsa/distilleR/issues}
}

}
\author{
\strong{Maintainer}: Luca Belmonte \email{luca.belmonte@efsa.europa.eu} (\href{https://orcid.org/0000-0002-7977-9170}{ORCID})

Authors:
\itemize{
  \item Lorenzo Copelli (\href{https://orcid.org/0009-0002-4305-065X}{ORCID})
  \item Fulvio Barizzone (\href{https://orcid.org/0009-0006-3035-520X}{ORCID})
  \item Dayana Stephanie Buzle (\href{https://orcid.org/0009-0003-2990-7431}{ORCID})
  \item Rafael Vieira (\href{https://orcid.org/0009-0009-0289-5438}{ORCID})
}

}
\keyword{internal}
