% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAsyncReportResult.R
\name{getAsyncReportResult}
\alias{getAsyncReportResult}
\title{Get the result of an asynchronous job to retrieve a Distiller report.}
\usage{
getAsyncReportResult(
  jobToken,
  format = c("excel", "csv"),
  distillerAsyncInstanceUrl = Sys.getenv("DISTILLER_ASYNC_INSTANCE_URL"),
  timeout = 1800
)
}
\arguments{
\item{jobToken}{\code{character} (string). The token associated to the successful
asynchronous job.}

\item{format}{\code{character} (string). The desired format for the document. It
can be either excel or csv.}

\item{distillerAsyncInstanceUrl}{\code{character} (string, optional). The
asynchronous Distiller instance URL.

By default: Sys.getenv("DISTILLER_ASYNC_INSTANCE_URL").}

\item{timeout}{\code{integer} (optional). The maximum number of seconds to wait
for the response.

By default: 1800 seconds (30 minutes).}
}
\value{
A data frame containing the Distiller report as designed within
DistillerSR.
}
\description{
This function gets the result of a successful Distiller asynchronous job to
retrieve a saved report associated with a given project ID. It requires a
valid asynchronous job token. The result is a dataframe containing metadata
about the saved report.
}
\examples{
\dontrun{
distillerToken_ <- getAuthenticationToken()

projects_ <- getProjects(distillerToken = distillerToken_)

reports_ <- getReports(
  projectId = projects_$id[1],
  distillerToken = distillerToken_)
  
job_ <- getReportAsync(
  projectId = projects_$id[1],
  reportID = reports_$id[7],
  format = "csv",
  distillerToken = distillerToken_)

jobToken_ <- job_$token

jobResult_ <- getAsyncReportResult(jobToken = jobToken_, format = "csv")
}

}
\seealso{
\code{\link{getAuthenticationToken}}

\code{\link{getProjects}}

\code{\link{getReports}}

\code{\link{getReportAsync}}

\code{\link{getAsyncReportStatus}}
}
