% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthenticationToken.R
\name{getAuthenticationToken}
\alias{getAuthenticationToken}
\title{Authenticate to a DistillerSR session.}
\usage{
getAuthenticationToken(
  distillerInstanceUrl = Sys.getenv("DISTILLER_INSTANCE_URL"),
  distillerKey = Sys.getenv("DISTILLER_API_KEY"),
  timeout = 1800
)
}
\arguments{
\item{distillerInstanceUrl}{\code{character} (string, optional). The base URL of
the DistillerSR instance.

By default: Sys.getenv("DISTILLER_INSTANCE_URL").}

\item{distillerKey}{\code{character} (string, optional). The personal access key
generated in DistillerSR.

By default: Sys.getenv("DISTILLER_API_KEY").}

\item{timeout}{\code{integer} (optional). The maximum number of seconds to wait
for the authentication response.

By default: 1800 seconds (30 minutes).}
}
\value{
A string containing a valid DistillerSR authentication token.
}
\description{
Authenticates a user to a DistillerSR instance using a personal access key.
The function returns a valid authentication token that can be used to access
protected DistillerSR API endpoints.
}
\details{
By default, the personal access key and the instance URL are read from the
environment variables \code{DISTILLER_API_KEY} and \code{DISTILLER_INSTANCE_URL}.
}
\examples{
\dontrun{
# If 'DISTILLER_INSTANCE_URL' and 'DISTILLER_API_KEY' are defined in your
# environment (e.g. .Renviron).
distillerToken_ <- getAuthenticationToken()

# If 'distillerInstanceUrl' and 'distillerKey' are to be specified manually.
distillerToken_ <- getAuthenticationToken(
  distillerInstanceUrl = "https://url.to.distiller.instance",
  distillerKey = "YOUR_API_KEY")
}

}
