% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportAsync.R
\name{getReportAsync}
\alias{getReportAsync}
\title{Submit an asynchronous job to retrieve a Distiller report.}
\usage{
getReportAsync(
  projectId,
  reportId,
  distillerInstanceUrl = Sys.getenv("DISTILLER_INSTANCE_URL"),
  distillerAsyncInstanceUrl = Sys.getenv("DISTILLER_ASYNC_INSTANCE_URL"),
  distillerToken,
  timeout = 1800
)
}
\arguments{
\item{projectId}{\code{integer}. The ID of the project as provided by DistillerSR.}

\item{reportId}{\code{integer}. The ID of the report as provided by DistillerSR.}

\item{distillerInstanceUrl}{\code{character} (string, optional). The Distiller
instance URL.

By default: Sys.getenv("DISTILLER_INSTANCE_URL").}

\item{distillerAsyncInstanceUrl}{\code{character} (string, optional). The
asynchronous Distiller instance URL.

By default: Sys.getenv("DISTILLER_ASYNC_INSTANCE_URL").}

\item{distillerToken}{\code{character} (string). The token the user gets once
authenticated.}

\item{timeout}{\code{integer} (optional). The maximum number of seconds to wait
for the response.

By default: 1800 seconds (30 minutes).}
}
\value{
A data frame containing metadata about the submitted job.
}
\description{
This function submits an asynchronous job to DistillerSR to retrieve a saved
report associated with a given project ID. It requires user authentication
and a valid asynchronous API endpoint URL. The result is a dataframe
containing metadata about the submitted job.
}
\examples{
\dontrun{
distillerToken_ <- getAuthenticationToken()

projects_ <- getProjects(distillerToken = distillerToken_)

reports_ <- getReports(
  projectId = projects_$id[1],
  distillerToken = distillerToken_)
  
job_ <- getReportAsync(
  projectId = projects_$id[1],
  reportID = reports_$id[7],
  format = "csv",
  distillerToken = distillerToken_)
}

}
\seealso{
\code{\link{getAuthenticationToken}}

\code{\link{getProjects}}

\code{\link{getReports}}

\code{\link{getAsyncReportStatus}}

\code{\link{getAsyncReportResult}}
}
