% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Make dataset xml node}
\usage{
make_dataset(.data, type)
}
\arguments{
\item{.data}{Pre-wrangled dataframe}

\item{type}{Type of data: hospital or emergency department}
}
\value{
XML node
}
\description{
Make dataset xml node
}
\examples{
data_right_vars <-
mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))

make_dataset(data_right_vars, "hosp")

}
