% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset_node.R
\name{make_dataset_node}
\alias{make_dataset_node}
\title{Make dataset xml node}
\usage{
make_dataset_node(data, content_group_id)
}
\arguments{
\item{data}{Pre-wrangled dataframe}

\item{content_group_id}{Code that identifies the content}
}
\value{
XML node
}
\description{
Make dataset xml node
}
\examples{
data_right_vars <-
  mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))

make_dataset_node(data_right_vars, "AS-HOSP")

}
