% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-make_header_node.R
\name{make_header_node}
\alias{make_header_node}
\title{Make Header node for XML Document}
\usage{
make_header_node(
  mcn,
  jurisdiction_code,
  content_group_id,
  submitter_email,
  submitter_name,
  submitter_title,
  state_fips_code
)
}
\arguments{
\item{mcn}{Metadata Control Number provided by EPHT.}

\item{jurisdiction_code}{Two-letter state abbreviation for the submitter
state.}

\item{content_group_id}{Code that identifies the content found in EPHT
documentation.}

\item{submitter_email}{Email of person submitting data to EPHT.}

\item{submitter_name}{First and last name of person submitting data to EPHT.}

\item{submitter_title}{Title of person submitting data to EPHT.}

\item{state_fips_code}{FIPS code of the submitter state.}
}
\value{
XML node
}
\description{
Create the Header Node of the XML document using the provided submission
metadata and tailor the XML format to match the supplied content group
identifier.
}
\examples{
make_header_node(
  mcn = "12345678-1234-1234-1234-123456789012",
  jurisdiction_code = "CA",
  content_group_id = "XX-HOSP",
  submitter_email = "myname@email.com",
  submitter_name = "First Last",
  submitter_title = "Data Analyst",
  state_fips_code = "06"
)
}
\seealso{
Other xml: 
\code{\link{make_dataset_node}()},
\code{\link{make_root_element}()},
\code{\link{make_xml_document}()}
}
\concept{xml}
