% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_months_worse}
\alias{make_months_worse}
\title{Convert integer month to string month with leading zero}
\usage{
make_months_worse(month_integer)
}
\arguments{
\item{month_integer}{Integer month}
}
\value{
String month with leading zero
}
\description{
Convert integer month to string month with leading zero
}
\examples{
make_months_worse(1)
}
\seealso{
Other helpers: 
\code{\link{collapse_ethnicity}()},
\code{\link{collapse_race}()},
\code{\link{parse_health_outcome_id}()}
}
\concept{helpers}
