% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_root_element.R
\name{make_root_element}
\alias{make_root_element}
\title{Make the root element: HospitalizationData or EmergencyDepartmentData}
\usage{
make_root_element(content_group_id)
}
\arguments{
\item{content_group_id}{Code that identifies the content}
}
\value{
XML node
}
\description{
Make the root element: HospitalizationData or EmergencyDepartmentData
}
\examples{
make_root_element("MI-HOSP")
}
