% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_health_outcome_id}
\alias{parse_health_outcome_id}
\title{Return the health outcome id based on the content group id}
\usage{
parse_health_outcome_id(content_group_id)
}
\arguments{
\item{content_group_id}{Code that identifies the content}
}
\value{
Integer health outcome id
}
\description{
Return the health outcome id based on the content group id
}
\examples{
parse_health_outcome_id("AS-HOSP")
}
