% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_helpers.R
\name{check_content_group_id}
\alias{check_content_group_id}
\title{Check the  validity of a Content Group Identifier}
\usage{
check_content_group_id(content_group_id)
}
\arguments{
\item{content_group_id}{Code that identifies the content}
}
\value{
list containing exit status code and success/failure message
}
\description{
Check the  validity of a Content Group Identifier
}
\examples{
check_content_group_id("AS-HOSP")
}
