% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_content.R
\name{check_data_content}
\alias{check_data_content}
\title{Check the validity of data content}
\usage{
check_data_content(data, content_group_id)
}
\arguments{
\item{data}{Pre-wrangled Dataframe.}

\item{content_group_id}{Code that identifies the content found in EPHT
documentation.}
}
\value{
list of exit statuses for each variable
}
\description{
There are finite values that are allowable in any one of the
variables that CDC want in a submission.  It is up to the user
to implement the correct logic as defined in the data dictionaries
and how-to-guides.

However if your logic is sound \code{check_data_content} will shoulder some
of the heavy lifting and provide users with its best guess about each
variable's formatting as to whether it is correct or not.  This is
a good way to quickly check your data before submitting it to the CDC.
This is \emph{not} a replacement for the CDC's EPHT Test Submission portal.
}
\section{Submission Check}{
If users set \code{check_first} = \code{TRUE} in \code{\link[=make_xml_document]{make_xml_document()}} or runs
\link{check_submission} or any of the other \code{check_* functions} then the a suite
of checks is run against the metadata, data structure and data content.
Please note that users do not need to run the whole suite of checks, they can
run each function piecemeal on their data as it is being prepared.

\link{check_submission} is called which is a wrapper around
the following functions:
\itemize{
\item \link{check_content_group_id}
\item \link{check_data}
\item \link{check_data_content}
\item \link{check_jurisdiction_code}
\item \link{check_mcn}
\item \link{check_state_fips_code}
\item \link{check_submitter_email}
\item \link{check_submitter_name}
\item \link{check_submitter_title}
}
}

\examples{
data <-
  mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))

check_data_content(data, "AS-HOSP")
}
