% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{collapse_ethnicity}
\alias{collapse_ethnicity}
\title{Collapse Ethnicity based on EPHT logic}
\usage{
collapse_ethnicity(ethnicity, na_is_unknown = TRUE)
}
\arguments{
\item{ethnicity}{Ethnicity string}

\item{na_is_unknown}{Convert NAs to "U" or leave as-is}
}
\value{
Ethnicity string or NA_character_
}
\description{
Collapse Ethnicity based on EPHT logic
}
\examples{
collapse_ethnicity("Hispanic")
}
