% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distiller-package.R
\docType{package}
\name{distiller-package}
\alias{distiller}
\alias{distiller-package}
\title{distiller: Distill EPHT Data Down to XML}
\description{
Centers for Disease Control and Prevention's (CDC) Environmental Public Health Tracking Network (EPHT) data submission process has a lot documentation. Much of it is buried and out of date, making it difficult for new users to decipher the exact specifications of the data submissions' XML. `distiller` aims to make it easier to created the desired XML.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asenetcky/distiller}
  \item Report bugs at \url{https://github.com/asenetcky/distiller/issues}
}

}
\author{
\strong{Maintainer}: Alexander Senetcky \email{asenetcky@gmail.com.com} (\href{https://orcid.org/0009-0009-3730-5397}{ORCID})

}
\keyword{internal}
