% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distiller-package.R
\docType{package}
\name{distiller-package}
\alias{distiller}
\alias{distiller-package}
\title{distiller: Distill Environmental Public Health Tracking Network Data Down to XML}
\description{
Convert dataframes of environmental public health tracking data into XML files that are compliant with the Centers for Disease Control and Prevention's (CDC) Environmental Public Health Tracking Network's (EPHT) data submission process. Users can also quickly validate their submission's data and metadata to reduce repeat visits to the time-consuming test submission portal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asenetcky/distiller}
  \item \url{https://asenetcky.github.io/distiller/}
  \item Report bugs at \url{https://github.com/asenetcky/distiller/issues}
}

}
\author{
\strong{Maintainer}: Alexander Senetcky \email{asenetcky@gmail.com} (\href{https://orcid.org/0009-0009-3730-5397}{ORCID}) [copyright holder]

}
\keyword{internal}
