% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjects.R
\name{getProjects}
\alias{getProjects}
\title{Get the list of the Distiller projects associated to the authenticated user.}
\usage{
getProjects(
  distillerInstanceUrl = Sys.getenv("DISTILLER_INSTANCE_URL"),
  distillerToken,
  timeout = 1800
)
}
\arguments{
\item{distillerInstanceUrl}{\code{character} (string). The distiller instance URL.

By default: Sys.getenv("DISTILLER_INSTANCE_URL").}

\item{distillerToken}{\code{character} (string). The token the user gets once
authenticated.}

\item{timeout}{\code{integer}. The maximum number of seconds to wait for the
response.

By default: 1800 seconds (30 minutes).}
}
\value{
A tibble with four columns:
\itemize{
\item \code{id}: The project ID.
\item \code{name}: The name of the project.
\item \code{de_project_id}.
\item \code{is_hidden}.
}
}
\description{
This function queries the DistillerSR API to retrieve the list of projects
accessible to the authenticated user. It requires an authentication token
and a valid API instance URL. The result is a dataframe listing available
projects.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
distillerToken_ <- getAuthenticationToken()

projects_ <- getProjects(distillerToken = distillerToken_)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{getAuthenticationToken}}
}
