% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{Get a Distiller report associated to a project of the authenticated user.}
\usage{
getReport(
  projectId,
  reportId,
  format = c("excel", "csv"),
  distillerInstanceUrl = Sys.getenv("DISTILLER_INSTANCE_URL"),
  distillerToken,
  timeout = 1800,
  attempts = 1,
  retryEach = 600,
  verbose = TRUE
)
}
\arguments{
\item{projectId}{\code{integer}. The ID of the project as provided by DistillerSR.}

\item{reportId}{\code{integer}. The ID of the report as provided by DistillerSR.}

\item{format}{\code{character} (string). The desired format for the document. It
can be either excel or csv.}

\item{distillerInstanceUrl}{\code{character} (string). The distiller instance URL.

By default: Sys.getenv("DISTILLER_INSTANCE_URL").}

\item{distillerToken}{\code{character} (string). The token the user gets once
authenticated.}

\item{timeout}{\code{integer}. The maximum number of seconds to wait for the
response.

By default: 1800 seconds (30 minutes).}

\item{attempts}{\code{integer}. The maximum number of attempts.

By default: 1 attempt.}

\item{retryEach}{\code{integer}. The delay between attempts.

By default: 600 seconds (10 minutes).}

\item{verbose}{\code{logical}. A flag to specify whether to make the function
verbose or not.

By default: TRUE.}
}
\value{
A data frame containing the Distiller report as designed within
DistillerSR.
}
\description{
This function queries the DistillerSR API to retrieve a saved report
associated with a given project ID. It requires user authentication
and a valid API endpoint URL. The result is a dataframe containing metadata
about the saved report.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
distillerToken_ <- getAuthenticationToken()
projects_ <- getProjects(distillerToken = distillerToken_)
reports_ <- getReports(
  projectId = projects_$id[1],
  distillerToken = distillerToken_)
  
report_ <- getReport(
  projectId = projects_$id[1],
  reportID = reports_$id[7],
  format = "csv",
  distillerToken = distillerToken_)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{getAuthenticationToken}}

\code{\link{getProjects}}

\code{\link{getReports}}
}
