% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReports.R
\name{getReports}
\alias{getReports}
\title{Get the list of the Distiller reports associated to a project of the
authenticated user.}
\usage{
getReports(
  projectId,
  distillerInstanceUrl = Sys.getenv("DISTILLER_INSTANCE_URL"),
  distillerToken,
  timeout = 1800
)
}
\arguments{
\item{projectId}{\code{integer}. The ID of the project as provided by DistillerSR.}

\item{distillerInstanceUrl}{\code{character} (string). The distiller instance URL.

By default: Sys.getenv("DISTILLER_INSTANCE_URL").}

\item{distillerToken}{\code{character} (string). The token the user gets once
authenticated.}

\item{timeout}{\code{integer}. The maximum number of seconds to wait for the
response.

By default: 1800 seconds (30 minutes).}
}
\value{
A tibble with four columns:
\itemize{
\item \code{id}: The ID of the saved report.
\item \code{name}: The name of the report.
\item \code{date}: The creation date of the report.
\item \code{view}: The format of the report (e.g., html, csv, excel).
}
}
\description{
This function queries the DistillerSR API to retrieve the list of saved
reports associated with a given project ID. It requires user authentication
and a valid API endpoint URL. The result is a dataframe containing metadata
about each saved report.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
distillerToken_ <- getAuthenticationToken()
projects_ <- getProjects(distillerToken = distillerToken_)

reports_ <- getReports(
  projectId = projects_$id[1],
  distillerToken = distillerToken_)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{getAuthenticationToken}}

\code{\link{getProjects}}
}
