% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make the root element}
\usage{
make_data(content_group_id)
}
\arguments{
\item{content_group_id}{Code that identifies the content}
}
\value{
XML node
}
\description{
Make the root element
}
\examples{
make_data("MI-HOSP")
}
