% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset_node.R
\name{make_dataset_node}
\alias{make_dataset_node}
\title{Make dataset xml node}
\usage{
make_dataset_node(data, content_group_id)
}
\arguments{
\item{data}{Pre-wrangled Dataframe.}

\item{content_group_id}{Code that identifies the content found in EPHT
documentation.}
}
\value{
XML node
}
\description{
Create the Dataset Node of the XML document using the provided data and
content group identifier.  The data variables are converted to row-based
XML child nodes and the elements ordered and named according to the content
group identifer.
}
\examples{
data_right_vars <-
  mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))

make_dataset_node(data_right_vars, "AS-HOSP")

}
\seealso{
Other xml: 
\code{\link{make_header_node}()},
\code{\link{make_root_element}()},
\code{\link{make_xml_document}()}
}
\concept{xml}
