% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_header_node.R
\name{make_header_node}
\alias{make_header_node}
\title{Make Header node for XML Document}
\usage{
make_header_node(
  mcn,
  jurisdiction_code,
  content_group_id,
  submitter_email,
  submitter_name,
  submitter_title,
  state_fips_code
)
}
\arguments{
\item{mcn}{Metadata Control Number provided by epht}

\item{jurisdiction_code}{Two-letter state abbreviation}

\item{content_group_id}{Code that identifies the content}

\item{submitter_email}{Email of person submitting data to epht}

\item{submitter_name}{First and last name of person submitting data}

\item{submitter_title}{Title of person submitting data}

\item{state_fips_code}{FIPS code of the state}
}
\value{
XML node
}
\description{
Make Header node for XML Document
}
\examples{
make_header_node(
  mcn = "12345678-1234-1234-1234-123456789012",
  jurisdiction_code = "CA",
  content_group_id = "XX-HOSP",
  submitter_email = "myname@email.com",
  submitter_name = "First Last",
  submitter_title = "Data Analyst",
  state_fips_code = "06"
)
}
