% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xml_document.R
\name{make_xml_document}
\alias{make_xml_document}
\title{Create an XML document}
\usage{
make_xml_document(
  data,
  content_group_id,
  mcn,
  jurisdiction_code,
  state_fips_code,
  submitter_email,
  submitter_name,
  submitter_title,
  check_first = FALSE
)
}
\arguments{
\item{data}{Pre-wrangled dataframe}

\item{content_group_id}{Code that identifies the content}

\item{mcn}{Metadata Control Number provided by epht}

\item{jurisdiction_code}{Two-letter state abbreviation}

\item{state_fips_code}{FIPS code of the state}

\item{submitter_email}{Email of person submitting data to epht}

\item{submitter_name}{First and last name of person submitting data}

\item{submitter_title}{Title of person submitting data}

\item{check_first}{Quickly checks the validity of your submission}
}
\value{
XML document
}
\description{
Create an XML document
}
\examples{
data <-
mtcars |>
  dplyr::rename(
    month = mpg,
    agegroup = cyl,
    county = disp,
    ethnicity = hp,
    health_outcome_id = drat,
    monthly_count = wt,
    race = qsec,
    sex = vs,
    year = am
  ) |>
  dplyr::select(-c(gear, carb))
content_group_id <- "AS-HOSP"
mcn <- "1234-1234-1234-1234-1234"
jurisdiction_code <- "two_letter_code"
state_fips_code <- "1234"
submitter_email <- "submitter@email.com"
submitter_name <- "Submitter Name"
submitter_title <- "Submitter Title"
doc <- make_xml_document(
 data,
 content_group_id,
 mcn,
 jurisdiction_code,
 state_fips_code,
 submitter_email,
 submitter_name,
 submitter_title
)
}
