% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_health_outcome_id}
\alias{parse_health_outcome_id}
\title{Return the health outcome id based on the content group id}
\usage{
parse_health_outcome_id(content_group_id)
}
\arguments{
\item{content_group_id}{Code that identifies the content found in EPHT
documentation.}
}
\value{
Integer health outcome identifier
}
\description{
The EPHT requires that each row node in the XML document have a health
outcome identifier.
}
\examples{
parse_health_outcome_id("AS-HOSP")
}
\seealso{
Other helpers: 
\code{\link{collapse_ethnicity}()},
\code{\link{collapse_race}()},
\code{\link{make_months_worse}()}
}
\concept{helpers}
